few fixes

+++ src/main/java/com/github/service/impl/CredentialService.java
@@ -1,20 +1,24 @@
 package com.github.service.impl;
 
+import com.github.dto.ConfirmationCredentialDto;
 import com.github.dto.CredentialDto;
 import com.github.entity.Credential;
 import com.github.entity.User;
 import com.github.exception.security.CredentialNotFoundException;
+import com.github.exception.user.WrongPasswordException;
 import com.github.mapper.IMapper;
 import com.github.repository.CredentialRepository;
 import com.github.repository.impl.RoleRepository;
 import com.github.service.ICredentialService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.HashSet;
 
+@Slf4j
 @Service
 @Transactional
 @AllArgsConstructor
@@ -37,28 +41,58 @@ public class CredentialService implements ICredentialService {
                 .setRoles(new HashSet<>(){{ add(roleRepository.findByName("USER")); }});
 
         credentialRepository.save(currentCredential);
+
+        log.info("Credential successfully created.");
     }
 
     @Override
+    @Transactional(readOnly = true)
     public CredentialDto readCredential(Long credentialId) {
         Credential currentCredential = credentialRepository.findById(credentialId)
-                .orElseThrow(() -> new CredentialNotFoundException("Credential exception"));
+                .orElseThrow(() -> new CredentialNotFoundException("There is no such account."));
+
+        log.info("Credential successfully founded.");
 
         return credentialMapper.toDto(currentCredential, CredentialDto.class);
     }
 
     @Override
-    public CredentialDto updateCredential(Long credentialId, CredentialDto credentialDto) {
-        credentialDto.setId(credentialId);
+    public void updateCredential(Long credentialId, ConfirmationCredentialDto credentialDto) {
 
-        Credential currentCredential = credentialMapper.toEntity(credentialDto, Credential.class);
+        Credential currentCredential = credentialRepository.findById(credentialId)
+                .orElseThrow(() -> new CredentialNotFoundException("There is no such account."));
 
-        return credentialMapper.toDto(credentialRepository.save(currentCredential), CredentialDto.class);
+        log.info("Credential successfully founded.");
+
+        String receivedPassword = passwordEncoder.encode(credentialDto.getPasswordConfirmation());
+
+        if (!receivedPassword.equals(currentCredential.getPassword()))
+            throw new WrongPasswordException("Password mismatch :(");
+
+        currentCredential.setEmail(credentialDto.getUsername())
+                .setPassword(passwordEncoder.encode(credentialDto.getPassword()));
+
+        credentialRepository.save(currentCredential);
+
+        log.info("Credential successfully updated.");
     }
 
     @Override
-    public void deleteCredential(Long credentialId) {
+    public void deleteCredential(Long credentialId, ConfirmationCredentialDto credentialDto) {
+
+        Credential currentCredential = credentialRepository.findById(credentialId)
+                .orElseThrow(() -> new CredentialNotFoundException("There is no such account."));
+
+        log.info("Credential successfully founded.");
+
+        String receivedPassword = passwordEncoder.encode(credentialDto.getPasswordConfirmation());
+
+        if (!receivedPassword.equals(currentCredential.getPassword()))
+            throw new WrongPasswordException("Password mismatch :(");
+
         credentialRepository.deleteById(credentialId);
+
+        log.info("Credential successfully deleted.");
     }
 
     @Override
